Creates a stage sandbox phase to the installer.
Inside stage sandbox we now add all the respective header search paths to the sandbox `HeaderStore` and each pod target's `HeaderStore`. This sets up the work for incremental pod installation because we would like to analyze each pod target's build settings to determine if it has changed; however, we previously only added the header search paths during installation which is too late in the pipline.